ChangeTempoBase in lib-builtin-effects
