+++ src/effects/ChangeTempo.h
@@ -17,12 +17,7 @@
 #ifndef __AUDACITY_EFFECT_CHANGETEMPO__
 #define __AUDACITY_EFFECT_CHANGETEMPO__
 
-#if USE_SBSMS
-#include "SBSMSBase.h"
-#endif
-
-#include "SoundTouchBase.h"
-#include "ShuttleAutomation.h"
+#include "ChangeTempoBase.h"
 #include "StatefulEffectUIServices.h"
 #include <wx/weakref.h>
 
@@ -30,54 +25,6 @@ class wxSlider;
 class wxCheckBox;
 class wxTextCtrl;
 class ShuttleGui;
-
-class ChangeTempoBase : public SoundTouchBase
-{
-public:
-   static inline ChangeTempoBase *
-   FetchParameters(ChangeTempoBase &e, EffectSettings &) { return &e; }
-   static const ComponentInterfaceSymbol Symbol;
-
-   ChangeTempoBase();
-   virtual ~ChangeTempoBase();
-
-   // ComponentInterface implementation
-
-   ComponentInterfaceSymbol GetSymbol() const override;
-   TranslatableString GetDescription() const override;
-   ManualPageID ManualPage() const override;
-
-   // EffectDefinitionInterface implementation
-
-   EffectType GetType() const override;
-   bool SupportsAutomation() const override;
-
-   // Effect implementation
-
-   bool Init() override;
-   bool CheckWhetherSkipEffect(const EffectSettings &settings) const override;
-   bool Process(EffectInstance &instance, EffectSettings &settings) override;
-   double CalcPreviewInputLength(
-      const EffectSettings &settings, double previewLength) const override;
-
-protected:
-   const EffectParameterMethods& Parameters() const override;
-
-   bool           mUseSBSMS;
-   double         m_PercentChange;  // percent change to apply to tempo
-                                    // -100% is meaningless, but sky's the upper limit
-   double         m_FromBPM;        // user-set beats-per-minute. Zero means not yet set.
-   double         m_ToBPM;          // Zero value means not yet set.
-   double         m_FromLength;     // starting length of selection
-   double         m_ToLength;       // target length of selection
-
-   bool m_bLoopDetect;
-
-static constexpr EffectParameter Percentage{ &ChangeTempoBase::m_PercentChange,
-   L"Percentage", 0.0,  -95.0,   3000.0,  1  };
-static constexpr EffectParameter UseSBSMS{ &ChangeTempoBase::mUseSBSMS,
-   L"SBSMS",     false, false,   true,    1  };
-};
 
 class EffectChangeTempo :
     public ChangeTempoBase,
